Separated processing params from lib
