+++ src/core/smoothing/fixed.rs
@@ -1,8 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright © 2021-2022 Elvin Chen
 
-use super::*;
 use nalgebra::*;
+
+use super::*;
 use crate::gyro_source::TimeQuat;
 
 #[derive(Clone)]
@@ -83,7 +84,7 @@ impl SmoothingAlgorithm for Fixed {
         hasher.finish()
     }
 
-    fn smooth(&mut self, quats: &TimeQuat, duration: f64, _params: &crate::BasicParams) -> TimeQuat {
+    fn smooth(&mut self, quats: &TimeQuat, duration: f64, _params: &ProcessingParams) -> TimeQuat {
         if quats.is_empty() || duration <= 0.0 { return quats.clone(); }
 
         const DEG2RAD: f64 = std::f64::consts::PI / 180.0;
