+++ src/core/smoothing/horizon.rs
@@ -1,8 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright © 2021-2022 Elvin Chen
 
-use super::*;
 use nalgebra::*;
+
+use super::*;
 use crate::gyro_source::TimeQuat;
 
 #[derive(Clone)]
@@ -57,7 +58,7 @@ impl SmoothingAlgorithm for HorizonLock {
         hasher.finish()
     }
 
-    fn smooth(&mut self, quats: &TimeQuat, duration: f64, _params: &crate::BasicParams) -> TimeQuat { // TODO Result<>?
+    fn smooth(&mut self, quats: &TimeQuat, duration: f64, _params: &crate::ProcessingParams) -> TimeQuat { // TODO Result<>?
         if quats.is_empty() || duration <= 0.0 { return quats.clone(); }
 
         let sample_rate: f64 = quats.len() as f64 / (duration / 1000.0);
