+++ src/core/smoothing/velocity_dampened_v1.rs
@@ -1,12 +1,11 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright © 2021-2022 Aphobius
 
+use nalgebra::*;
 use std::collections::BTreeMap;
 
 use super::*;
-use nalgebra::*;
-use crate::gyro_source::TimeQuat;
-use crate::Quat64;
+use crate::gyro_source::{Quat64, TimeQuat};
 
 #[derive(Clone)]
 pub struct VelocityDampened {
@@ -112,7 +111,7 @@ impl SmoothingAlgorithm for VelocityDampened {
         hasher.finish()
     }
 
-    fn smooth(&mut self, quats: &TimeQuat, duration: f64, params: &crate::BasicParams) -> TimeQuat { // TODO Result<>?
+    fn smooth(&mut self, quats: &TimeQuat, duration: f64, params: &ProcessingParams) -> TimeQuat { // TODO Result<>?
         if quats.is_empty() || duration <= 0.0 { return quats.clone(); }
 
         let sample_rate: f64 = quats.len() as f64 / (duration / 1000.0);
